🐿 split animal modification and creation into component and subcomponent
