+++ src/backend/api/animals/controller.js
@@ -6,7 +6,7 @@ const router = new Router()
 
 router.get('/', (request, response) => {
   findAllAnimals()
-    .then(species => response.status(200).json(species))
+    .then(animals => response.status(200).json(animals))
     .catch((error) => {
       console.log(error) // eslint-disable-line
       response.status(500).send()
@@ -15,7 +15,7 @@ router.get('/', (request, response) => {
 
 router.get('/:id', (request, response) => {
   findById(request.params.id)
-    .then(species => response.status(200).json(species))
+    .then(animal => response.status(200).json(animal))
     .catch((error) => {
       console.log(error) // eslint-disable-line
       response.status(500).send()
@@ -24,12 +24,6 @@ router.get('/:id', (request, response) => {
 
 router.post('/new', (request, response) => {
   console.log(response.body)
-  // createNewSpecies(request.body)
-  //   .then(entry => response.status(200).json(entry))
-  //   .catch((error) => {
-  //     console.log(error) // eslint-disable-line
-  //     response.status(500).send()
-  //   })
 })
 
 export default router
