+++ src/frontend/animals/modification/index.js
@@ -1,0 +1,1 @@
+export { default } from './ModificationForm'
